oran: add WithTemplateParameter method
This PR adds a method to set keys in the TemplateParameters. TemplateParameters is a RawExtension but the method handles this transparently by adding two new unexported methods to set and get the field as a map[string]any. Since the schema of this field is defined by the ClusterTemplate, map[string]any is the most specific type and is what the operator itself uses internally.